Fix: cont. + cleaning + changelogs
